+++ gym/src/main/java/com/gym/gym/controller/HomeController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.CookieValue;
@@ -174,105 +175,6 @@ public String login(@CookieValue(value = "remember-id", required = false) Cookie
     //     return "error/403";
     // }
     
-
-    @GetMapping("user/findId")
-    public String findId() {
-        return "/user/findId";
-    }
-
-    @PostMapping("/user/findId")
-    public String findId(Model model, @RequestParam("name") String name,
-                                      @RequestParam("phone") String phone, @RequestParam("question") String question,
-                                      @RequestParam("answer") String answer) throws Exception {
-        
-                                log.info(name + "이름");
-                                log.info(phone + "전화번호");
-                                log.info(question + "질문");
-                                log.info(answer + "답변");
-        // 이름, 전화번호, 질문, 답변을 기준으로 사용자 찾기
-        Users foundUser = userService.findUserByDetails(name, phone, question, answer);
-        
-        log.info(foundUser + "아이디찾기");
-        if (foundUser != null && foundUser.getId() != null) {
-            model.addAttribute("user", foundUser);
-            model.addAttribute("no", 1);
-            return "/user/find";
-        } else {
-            model.addAttribute("users", null);
-            model.addAttribute("message", "사용자를 찾을 수 없습니다.");
-            return "/user/find";
-        }
-    }
-    
-
-    @GetMapping("/user/findPassword")
-    public String findPassword() {
-        return "/user/findPassword";
-    }
-
-    @PostMapping("/user/findPassword")
-    public String findPassword(Model model, @RequestParam("name") String name,
-                                      @RequestParam("phone") String phone, @RequestParam("question") String question,
-                                      @RequestParam("answer") String answer, @RequestParam("id") String id) throws Exception {
-        
-                                log.info(name + "이름");
-                                log.info(phone + "전화번호");
-                                log.info(question + "질문");
-                                log.info(answer + "답변");
-        // 이름, 전화번호, 질문, 답변을 기준으로 사용자 찾기
-        Users foundUser = userService.findUserByPassword(name, phone, question, answer, id);
-        
-        log.info(foundUser + "비밀번호찾기");
-        if (foundUser != null && foundUser.getId() != null) {
-            String code = UUID.randomUUID().toString().substring(0, 6) ;
-
-            foundUser.setCode(code);
-           model.addAttribute("code", code);
-           model.addAttribute("no", foundUser.getNo());
-            userService.codeInsert(foundUser);
-            return "/user/changePassword";
-        } else {
-            model.addAttribute("users", null);
-            model.addAttribute("message", "사용자를 찾을 수 없습니다.");
-            return "/user/find";
-        }
-
-        
-    }
-    
-@GetMapping("/user/changePassword")
-public String changePassword() {
-    return "/user/changePassword";
-}
-
-@PostMapping("/user/changePassword")
-public String changePassword(@RequestParam("code") String code, @RequestParam("password") String password, 
-                            @RequestParam("newPassword") String newPassword, @RequestParam("no") Long no) throws Exception {
-    Users user = userService.select(no);
-      
-    if (user.getPassword() == password ) {
-        user.setPassword(newPassword);
-      int result = userService.passwordUpdate(user);
-        if (result > 0) {
-            user.setCode(null);
-            userService.update(user);
-            return "/login";
-        }
-        else{
-            user.setCode(null);
-            userService.update(user);
-            return "/login";
-
-        }
-
-        
-    }
-    user.setCode(null);
-    userService.update(user);
-    return "/login";
-}
-
-
