+++ gym/src/main/resources/com/gym/gym/mapper/UserMapper.xml
@@ -116,12 +116,14 @@
     </select>
 
 
-     <insert id="codeInsert">
-    INSERT INTO users (code)
-    SET (code = #{code})
-    </insert>
+     <update id="codeInsert">
+    UPDATE users
+    SET code = #{code}
+      WHERE no = #{no}
+    </update>
 
     
+
 
     <update id="passwordUpdate">
     UPDATE users
