+++ gym/src/main/resources/templates/join.html
@@ -11,7 +11,6 @@
 </head>
 <body>
  
-  
     <div class="container col-12 col-lg-4">
         <div class="px-4 py-5 mt-5 text-center">
             <h1 class="display-5 fw-bold text-body-emphasis">회원 가입</h1>
@@ -25,11 +24,11 @@
                 
                 <div class="input-group my-2">
                     <div class="form-floating" id="box-id">
-                        <input type="text" class="form-control" id="id" name="id" value="" placeholder="아이디" autofocus>
+                        <input type="text" class="form-control" id="id" name="id" placeholder="아이디" autofocus>
                         <label for="username">아이디</label>
                     </div>
                     <div class="input-group-append">
-                        <button type="button" class="btn btn-lg btn-outline-secondary h-100" onclick="checkId()" >중복확인</button>
+                        <button type="button" class="btn btn-lg btn-outline-secondary h-100" onclick="checkId()">중복확인</button>
                     </div>
                 </div>
                 <div class="form-floating my-2">
@@ -41,37 +40,35 @@
                     <label for="passwordCheck">비밀번호 확인</label>
                 </div>
                 <div class="form-floating my-2">
-                    <input type="text" class="form-control" id="name" name="name" value="" placeholder="이름"
-                        autofocus="">
+                    <input type="text" class="form-control" id="name" name="name" placeholder="이름">
                     <label for="name">이름</label>
                 </div>
                 <div class="form-floating my-2">
-                    <input type="text" class="form-control" id="birth" name="birth" placeholder="연락처를 입력해주세요.">
-                    <label for="passwordCheck">생일</label>
+                    <input type="text" class="form-control" id="birth" name="birth" placeholder="생일">
+                    <label for="birth">생일</label>
                 </div>
                 <div class="form-floating my-2">
-                    <input type="text" class="form-control" id="gender" name="gender">
-                    <label for="passwordCheck">성별</label>
+                    <input type="text" class="form-control" id="gender" name="gender" placeholder="성별">
+                    <label for="gender">성별</label>
                 </div>
 
                 <div class="form-floating my-2">
-                    <input type="text" class="form-control" id="email" name="email" value="" placeholder="이메을"
-                        autofocus="">
+                    <input type="text" class="form-control" id="email" name="email" placeholder="이메일을 입력해주세요" autofocus="">
                     <label for="email">이메일</label>
                 </div>
                 <div class="form-floating my-2">
                     <input type="text" class="form-control" id="phone" name="phone" placeholder="연락처를 입력해주세요.">
-                    <label for="passwordCheck">연락처</label>
+                    <label for="phone">연락처</label>
                 </div>
 
                 <div class="form-floating my-2">
-                    <input type="text" class="form-control" id="question" name="question" value="강아지이름은?">
-                    <label for="passwordCheck">질문</label>
+                    <input type="text" class="form-control" id="question" name="question" value="강아지 이름은?" placeholder="질문">
+                    <label for="question">질문</label>
                 </div>
                 
                 <div class="form-floating my-2">
-                    <input type="text" class="form-control" id="answer" name="answer" placeholder="연락처를 입력해주세요.">
-                    <label for="passwordCheck">답변</label>
+                    <input type="text" class="form-control" id="answer" name="answer" placeholder="답변">
+                    <label for="answer">답변</label>
                 </div>
                 
                 <div class="d-grid gap-2">
@@ -86,10 +83,7 @@
     <!-- bootstrap js -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"></script>
 
-
     <script>
-
-
         // 💍 CRSF TOKEN
         const csrfToken = "[[${_csrf.token}]]"
 
@@ -106,10 +100,6 @@
             }
 
             try {
-                // 아이디 중복 확인
-                // fetch(URL, 정보)
-                // - 정보 : method, headers(헤더)
-                
                 const response = await fetch(`/check/${username}`, {
                     method: 'GET',
                     headers: {
@@ -142,24 +132,40 @@
             }
         }
 
-
         /*
             제출 확인
-            - 아이디 중복 체크
+            - 아이디 중복 체크, 비밀번호 확인, 빈 칸 검사
         */
         async function checkSubmit(event) {
-            event.preventDefault();                         // 폼 제출 방지
-            
+            event.preventDefault();  // 폼 제출 방지
+
             // 아이디 중복 체크
             const isIdAvailable = await checkId();
             if (!isIdAvailable) {
                 return;
             }
 
-            document.getElementById("form").submit();   
+            // 비밀번호 확인
+            const password = document.getElementById("password").value;
+            const passwordCheck = document.getElementById("passwordCheck").value;
+            if (password !== passwordCheck) {
+                alert("비밀번호가 일치하지 않습니다.");
+                return;
+            }
+
+            // 필수 입력 항목 체크
+            const requiredFields = ['id', 'password', 'passwordCheck', 'name', 'birth', 'gender', 'email', 'phone', 'question', 'answer'];
+            for (let field of requiredFields) {
+                const fieldValue = document.getElementById(field).value;
+                if (!fieldValue) {
+                    alert(`${document.querySelector(`[for=${field}]`).textContent}을(를) 입력해주세요.`);
+                    return;
+                }
+            }
+
+            // 모든 유효성 검사 통과 시 폼 제출
+            document.getElementById("form").submit();
         }
-
-
     </script>
 </body>
 </html>
