+++ gym/src/main/resources/templates/login.html
@@ -10,7 +10,9 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-   
+    <script th:if="${error != null}">
+        alert('비밀번호 변경에 실패했습니다. 비밀번호를 확인해 주세요.');
+    </script>
     
     <div class="container col-12 col-md-6 col-lg-4">
 
@@ -70,7 +72,7 @@
             </form>
         </main>
     </div>
-
+ 
     <!-- bootstrap -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"></script>
 </body>
