+++ gym/src/main/resources/templates/user/changePassword.html
@@ -7,11 +7,32 @@
 </head>
 <body>
     <h1>비밀번호 바꾸기</h1>
-    <form action="user/"></form>
-    <input type="password" name="" id="">
-    <input type="password" name="" id="">
-    <input type="password" name="" id="">
-    <input type="password" name="" id="">
+    <form action="changePassword" method="post" onsubmit="return checkSubmit(event)">
+        <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
+        <input type="hidden" name="code" th:value="${code}">
+          <input type="hidden" name="no" th:value="${no}">
+
+    <label >현재 비밀번호</label> <input type="password" name="password" id="">
+    <label> 새로운 비밀번호 </label> <input type="password" name="newPassword" id="">
+    <button type="submit"> 수정하기</button>
+    </form>
+<script>
+ async function checkSubmit(event) {
+            event.preventDefault();  // 폼 제출 방지
+
+
+            // 비밀번호 확인
+            const password = document.getElementById("password").value;
+            const passwordCheck = document.getElementById("passwordCheck").value;
+            if (password !== passwordCheck) {
+                alert("비밀번호가 일치하지 않습니다.");
+                return;
+            }
+
+            document.getElementById("form").submit();
+        }
+
+</script>
 
 </body>
 </html>
