+++ gym/src/main/resources/templates/user/find.html
@@ -8,7 +8,7 @@
 <body>
     
     <h3>게정찾기</h3>
-    <div th:if="${no == 1}">
+   
         <h1>아이디 찾기 결과</h1>
         <div th:if="${user != null}">
             <p>사용자 이름: <span th:text="${user.name}"></span></p>
@@ -17,18 +17,7 @@
         <div th:if="${user == null}">
             <p th:text="${message}">사용자를 찾을 수 없습니다.</p>
         </div>
-    </div>
-    <div th:if="${no == 2}">
-        <h1>비밀번호 찾기 결과</h1>
-        <div th:if="${user != null}">
-            <p>사용자 이름: <span th:text="${user.name}"></span></p>
-            <p>사용자 아이디: <span th:text="${user.id}"></span></p>
-            <p>사용자 비밀번호: <span th:text="${user.password}"></span></p>
-        </div>
-        <div th:if="${user == null}">
-            <p th:text="${message}">사용자를 찾을 수 없습니다.</p>
-        </div>
-    </div>
+   
  
 </body>
 </html>
