+++ gym/src/main/resources/templates/user/findId.html
@@ -4,19 +4,52 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>아이디 찾기</title>
+    
 </head>
 <body>
-    <h1> 아이디 찾기</h1>
-     <form action="findId" method="post">
+    <h1>아이디 찾기</h1>
+    <form action="findId" method="post" onsubmit="return validateForm(event)">
         <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
-    <span>이름 : </span><input type="text" name="name" id="">
-    <span>전화번호 :  </span><input type="text" name="phone" id="">
-    <h5> 질문 : </h5>
-    <select name="question" id="">
-      <option  name="question" value="강아지 이름은?" id="">강아지 이름은?</option>
-    </select>
-    <h5>답변 : </h5> <input type="text" name="answer" >
-    <button type="submit" >찾기</button>
-</form>
+
+        <div>
+            <span>이름 : </span>
+            <input type="text" name="name">
+        </div>
+        <div>
+            <span>전화번호 : </span>
+            <input type="text" name="phone">
+        </div>
+        <div>
+            <h5>질문 : </h5>
+            <select name="question">
+                <option value="강아지 이름은?">강아지 이름은?</option>
+            </select>
+        </div>
+        <div>
+            <h5>답변 : </h5>
+            <input type="text" name="answer">
+        </div>
+        <div>
+            <button type="submit">찾기</button>
+        </div>
+    </form>
+
+    <script>
+        function validateForm(event) {
+            const requiredFields = ['name', 'phone', 'question', 'answer'];
+
+            for (let field of requiredFields) {
+                const inputElement = document.getElementsByName(field)[0];
+                const labelText = inputElement.previousElementSibling.textContent.trim();
+                const value = inputElement.value.trim();
+
+                if (!value) {
+                    alert(`${labelText}을(를) 입력해주세요.`);
+                    return false;
+                }
+            }
+            return true; 
+        }
+    </script>
 </body>
 </html>
