+++ gym/src/main/resources/templates/user/findPassword.html
@@ -7,17 +7,53 @@
 </head>
 <body>
     <h1> 비밀번호 찾기</h1>
-     <form action="changePassword" method="post">
+    <form action="findPassword" method="post" onsubmit="return validateForm()">
         <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
-    <span>이름 : </span><input type="text" name="name" id="">
-    <span>아이디 : </span><input type="text" name="id" id="">
-    <span>전화번호 :  </span><input type="text" name="phone" id="">
-    <h5> 질문 : </h5>
-    <select name="question" id="">
-      <option  name="question" value="강아지 이름은?" id="">강아지 이름은?</option>
-    </select>
-    <h5>답변 : </h5> <input type="text" name="answer" >
-    <button type="submit" >찾기</button>
-</form>
+        
+        <span>이름 :</span>
+        <input type="text" name="name" id="name">
+        <br>
+        <span>아이디 :</span>
+        <input type="text" name="id" id="id">
+        <br>
+        <span>전화번호 :</span>
+        <input type="text" name="phone" id="phone">
+        <br>
+        <h5>질문 :</h5>
+        <select name="question" id="question">
+            <option value="">질문을 선택해주세요</option>
+            <option value="강아지 이름은?">강아지 이름은?</option>
+        </select>
+        <br>
+        <span>답변 :</span>
+        <input type="text" name="answer" id="answer">
+        <br>
+        <button type="submit">찾기</button>
+    </form>
+    
+    <script>
+        function validateForm() {
+            // 입력 필드를 배열로 정의
+            const fields = [
+                { id: "name", label: "이름" },
+                { id: "id", label: "아이디" },
+                { id: "phone", label: "전화번호" },
+                { id: "question", label: "질문" },
+                { id: "answer", label: "답변" },
+            ];
+
+            // 모든 필드 확인
+            for (let field of fields) {
+                const input = document.getElementById(field.id);
+                if (!input.value.trim()) { // 값이 비어있으면 경고창
+                    alert(`${field.label}을(를) 입력해주세요.`);
+                    return false; // 폼 제출 중단
+                }
+            }
+
+            // 모든 필드가 채워졌을 경우 폼 제출
+            return true;
+        }
+    </script>
 </body>
 </html>
